src/requirements.txt
requests>=2.31.0
PyYAML>=6.0
python-dateutil>=2.8.2
pytz>=2023.3
openpyxl>=3.1.0
pandas>=2.0.0
PySide6>=6.6.0
python-docx>=1.1.0
pdfminer.six>=20221105
langdetect>=1.0.9
vaderSentiment>=3.3.2
sentence-transformers>=2.0
scikit-learn>=1.3.0
textstat>=0.7.0
pydantic>=2.0.0
spacy>=3.7.0
# Optional: only needed if whisper.cpp is not installed
# faster-whisper>=1.0.0